requirements.txt
# Core Databricks
databricks-sdk>=0.27.0
databricks-sql-connector>=3.0.0
databricks-vectorsearch>=0.22

# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-databricks>=0.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Synthetic Data Generation
faker>=20.0.0

# Streamlit (Dashboard)
streamlit>=1.31.0
plotly>=5.18.0
altair>=5.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tenacity>=8.2.0
